+++ package.json
@@ -14,6 +14,7 @@
     "browserify-zlib": "^0.2.0",
     "cosmjs-types": "^0.5.1",
     "crypto-browserify": "^3.12.0",
+    "graz": "^0.0.27",
     "path-browserify": "^1.0.1",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
